Enable ruff comprehension and ruff-specific rules
Add flake8-comprehensions (C4) and the ruff-specific defect rules (RUF)
to the lint selection and fix the 22 reported findings: redundant
list()/dict() calls, int() wrappers around math.ceil() / round() which
already return integers, list-comprehension arguments passed to
all(), list concatenations rewritten to unpacking (behavior unchanged),
an unused unpacked variable, __all__ sorted, and a ClassVar annotation
for the read-only alignment expectation in the IBM test. The F401 noqa
in lpu/__init__.py documents the intentional availability probe import
and stays.

+++ lpu/commands/exec_parallel.py
@@ -172,11 +172,11 @@ def splitFile(conf: Config) -> bool | None:
     prefix = getSplitPrefix(conf)
     if not splitSize:
         #splitSize = int( math.ceil(float(lineCount) / threads) )
-        splitSize = conf.data.splitSize = int( math.ceil(float(lineCount) / numChunks) )
+        splitSize = conf.data.splitSize = math.ceil(float(lineCount) / numChunks)
         logger.info(f'Split size: {splitSize}')
     #splitCount = conf.data.splitCount = int(math.ceil(float(lineCount) / splitSize))
     #splitCount = int(math.ceil(float(lineCount) / splitSize))
-    numChunks = conf.data.numChunks = int(math.ceil(float(lineCount) / splitSize))
+    numChunks = conf.data.numChunks = math.ceil(float(lineCount) / splitSize)
     #digits = conf.data.digits = len(str(splitCount))
     digits = conf.data.digits = len(str(numChunks))
     logger.info(f'Splitting into: "{prefix}.*"')
